requirements.txt

streamlit
PyPDF2
python-docx
transformers
sentence-transformers
faiss-cpu  # Use faiss-gpu if you have a GPU
numpy
